feat(frontend): add Help & Support page with sidebar navigation
Add a dedicated Help & Support experience to BirdNET-Go's frontend:

- New Help page at /ui/help with 4 cards: Report a Bug (guided workflow),
  Ask a Question, Diagnostics (embedded support dump), Quick Links
- Collapsible Help sidebar section between System and Settings with
  internal Help & Support link plus external Report Bug and Ask a Question
- Report Bug and Ask a Question links added to header settings menu
- SupportDumpCard extracted as shared component from SupportSettingsPage,
  used by both Help page and Settings > Support
- Shared externalUrls.ts module centralizes GitHub URLs across 3 files
- Full i18n: 15 new keys with native translations in all 15 locales

Bonus fixes for pre-existing issues in touched files:
- SupportDumpCard: unique HTML IDs, timer cleanup on unmount, onMount
  instead of $effect for one-time fetch
- DesktopSidebar: hardcoded aria-labels replaced with t() calls,
  hardcoded GitHub URL replaced with shared constant
- SupportSettingsPage: hardcoded English aria-label replaced with t()

Closes Forgejo #689
