+++ frontend/src/lib/desktop/components/ui/HeaderSettingsMenu.svelte
@@ -12,9 +12,23 @@
   import { getLogger } from '$lib/utils/logger';
   import { resetDateToToday } from '$lib/utils/datePersistence';
   import { clearGuestLayout } from '$lib/stores/guestDashboardLayout';
-  import { Settings, Sun, Moon, Pencil, RotateCcw } from '@lucide/svelte';
+  import {
+    Settings,
+    Sun,
+    Moon,
+    Pencil,
+    RotateCcw,
+    Bug,
+    MessageCircleQuestion,
+    ExternalLink,
+  } from '@lucide/svelte';
   import GithubIcon from '$lib/desktop/components/ui/GithubIcon.svelte';
   import { dropdown } from '$lib/utils/transitions';
+  import {
+    GITHUB_REPO_URL,
+    GITHUB_ISSUES_URL,
+    GITHUB_DISCUSSIONS_URL,
+  } from '$lib/utils/externalUrls';
   import ConfirmModal from '$lib/desktop/components/modals/ConfirmModal.svelte';
 
   const logger = getLogger('dashboard');
@@ -224,9 +238,35 @@
         <!-- Divider -->
         <div class="my-1 border-t border-[var(--color-base-content)]/10" aria-hidden="true"></div>
 
+        <!-- Report Bug -->
+        <a
+          href={GITHUB_ISSUES_URL}
+          target="_blank"
+          rel="noopener noreferrer"
+          class="flex w-full items-center gap-3 rounded-md px-3 py-2 text-sm font-normal text-[var(--color-base-content)] transition-colors duration-150 hover:bg-[var(--color-base-content)]/10"
+          aria-label={t('navigation.reportBugAriaLabel')}
+        >
+          <Bug class="size-4 shrink-0 text-[var(--color-base-content)]/70" />
+          <span class="flex-1">{t('navigation.reportBug')}</span>
+          <ExternalLink class="size-3 text-[var(--color-base-content)]/40" />
+        </a>
+
+        <!-- Ask a Question -->
+        <a
+          href={GITHUB_DISCUSSIONS_URL}
+          target="_blank"
+          rel="noopener noreferrer"
+          class="flex w-full items-center gap-3 rounded-md px-3 py-2 text-sm font-normal text-[var(--color-base-content)] transition-colors duration-150 hover:bg-[var(--color-base-content)]/10"
+          aria-label={t('navigation.askQuestionAriaLabel')}
+        >
+          <MessageCircleQuestion class="size-4 shrink-0 text-[var(--color-base-content)]/70" />
+          <span class="flex-1">{t('navigation.askQuestion')}</span>
+          <ExternalLink class="size-3 text-[var(--color-base-content)]/40" />
+        </a>
+
         <!-- GitHub link -->
         <a
-          href="https://github.com/tphakala/birdnet-go"
+          href={GITHUB_REPO_URL}
           target="_blank"
           rel="noopener noreferrer"
           class="flex w-full items-center gap-3 rounded-md px-3 py-2 text-sm font-normal text-[var(--color-base-content)] transition-colors duration-150 hover:bg-[var(--color-base-content)]/10"
