+++ frontend/src/lib/utils/externalUrls.ts
@@ -1,0 +1,4 @@
+export const GITHUB_REPO_URL = 'https://github.com/tphakala/birdnet-go';
+export const GITHUB_ISSUES_URL = 'https://github.com/tphakala/birdnet-go/issues/new/choose';
+export const GITHUB_DISCUSSIONS_URL = 'https://github.com/tphakala/birdnet-go/discussions';
+export const GITHUB_RELEASES_URL = 'https://github.com/tphakala/birdnet-go/releases';
