+++ frontend/static/messages/pt.json
@@ -228,7 +228,15 @@
     "analyticsSubmenu": "Submenu de análises",
     "settingsSubmenu": "Submenu de configurações",
     "systemSubmenu": "Submenu do sistema",
-    "systemTerminal": "Terminal"
+    "systemTerminal": "Terminal",
+    "help": "Ajuda",
+    "helpSubmenu": "Submenu de ajuda",
+    "helpAndSupport": "Ajuda e suporte",
+    "reportBug": "Reportar erro",
+    "askQuestion": "Fazer uma pergunta",
+    "reportBugAriaLabel": "Reportar um erro no GitHub (abre em nova janela)",
+    "askQuestionAriaLabel": "Fazer uma pergunta no GitHub Discussions (abre em nova janela)",
+    "viewOnGithubAriaLabel": "Ver BirdNET-Go no GitHub (abre em nova janela)"
   },
   "about": {
     "title": "BirdNET-Go",
@@ -4693,5 +4701,22 @@
     "bannerMessage": "As alterações de configuração requerem uma reinicialização:",
     "bannerAction": "Reiniciar agora",
     "restartFailed": "Falha ao iniciar a reinicialização"
+  },
+  "help": {
+    "title": "Ajuda e suporte",
+    "subtitle": "Obtenha ajuda, reporte problemas e conecte-se com a comunidade",
+    "reportBug": {
+      "description": "Se encontrar um problema, primeiro gere um pacote de suporte para capturar as informações do sistema e os registros, depois abra um relatório de erro no GitHub."
+    },
+    "askQuestion": {
+      "description": "Tem uma pergunta sobre configuração, funcionalidades ou utilização? O fórum GitHub Discussions é o melhor local para obter ajuda da comunidade."
+    },
+    "diagnostics": {
+      "description": "Gere um pacote de suporte para ajudar a diagnosticar problemas. Captura a configuração do sistema e os registros recentes."
+    },
+    "quickLinks": {
+      "title": "Links rápidos",
+      "releases": "Notas de lançamento"
+    }
   }
 }
