+++ frontend/static/messages/sv.json
@@ -228,7 +228,15 @@
     "analyticsSubmenu": "Undermeny för analys",
     "settingsSubmenu": "Undermeny för inställningar",
     "systemSubmenu": "Undermeny för system",
-    "systemTerminal": "Terminal"
+    "systemTerminal": "Terminal",
+    "help": "Hjälp",
+    "helpSubmenu": "Hjälp-undermeny",
+    "helpAndSupport": "Hjälp och support",
+    "reportBug": "Rapportera bugg",
+    "askQuestion": "Ställ en fråga",
+    "reportBugAriaLabel": "Rapportera en bugg på GitHub (öppnas i nytt fönster)",
+    "askQuestionAriaLabel": "Ställ en fråga på GitHub Discussions (öppnas i nytt fönster)",
+    "viewOnGithubAriaLabel": "Visa BirdNET-Go på GitHub (öppnas i nytt fönster)"
   },
   "about": {
     "title": "BirdNET-Go",
@@ -4693,5 +4701,22 @@
     "bannerMessage": "Konfigurationsändringar kräver en omstart:",
     "bannerAction": "Starta om nu",
     "restartFailed": "Det gick inte att starta om"
+  },
+  "help": {
+    "title": "Hjälp och support",
+    "subtitle": "Få hjälp, rapportera problem och anslut till gemenskapen",
+    "reportBug": {
+      "description": "Om du stöter på ett problem, generera först ett supportpaket för att samla systeminformation och loggar, öppna sedan en felrapport på GitHub."
+    },
+    "askQuestion": {
+      "description": "Har du en fråga om installation, funktioner eller användning? GitHub Discussions är det bästa stället att få hjälp från gemenskapen."
+    },
+    "diagnostics": {
+      "description": "Generera ett supportpaket för att hjälpa till med felsökning. Det samlar din systemkonfiguration och senaste loggar."
+    },
+    "quickLinks": {
+      "title": "Snabblänkar",
+      "releases": "Versionsanteckningar"
+    }
   }
 }
